- casi acabado

+++ src/DBControl/Eliminar.java
@@ -10,35 +10,58 @@ import org.hibernate.Session;
  */
 public class Eliminar {
 
-    public static void libroPorId(int Id){
+    public static boolean libroPorId(int id){
 
         Session session = Acceso.getSession("Eliminar_libroPorId");
 
         session.beginTransaction();
-        Libros libro = new Libros(Id);
-        session.delete(libro);
-        session.getTransaction().commit();
+        Libros libro = DBControl.Consultar.LibroPorId(id);
+
+        if (libro.getId() == 0) return false;
+        else{
+            session.delete(libro);
+            session.getTransaction().commit();
+            return true;
+        }
     }
 
-    public static void socioPorId(int Id){
+    public static boolean socioPorId(int id){
 
         Session session = Acceso.getSession("Eliminar_SocioPorId");
 
         session.beginTransaction();
-        Socios socio = new Socios(Id);
-        session.delete(socio);
-        session.getTransaction().commit();
+        Socios socio = DBControl.Consultar.SocioPorId(id);
+
+        if (socio.getId() == 0) return false;
+        else {
+            session.delete(socio);
+            session.getTransaction().commit();
+            return true;
+        }
     }
 
-    public static void PrestamosPorId(int Id){
+    public static boolean prestamoPorId(int id){
 
         Session session = Acceso.getSession("Eliminar_PrestamosPorId");
 
         session.beginTransaction();
-        Prestamo prestamo = new Prestamo(Id);
-        session.delete(prestamo);
-        session.getTransaction().commit();
+        Prestamo prestamo = DBControl.Consultar.PrestamoPorId(id);
+
+        if (prestamo.getId() == 0) return false;
+        else {
+            session.delete(prestamo);
+            session.getTransaction().commit();
+            return true;
+        }
     }
 
+    public static void Prestamos(Prestamo prestamo){
+        Session session = Acceso.getSession("Eliminar_Prestamo");
+
+        session.beginTransaction();
+        session.delete(prestamo);
+        session.getTransaction().commit();
+
+    }
 
 }
